+++ sdmx-dl-api/src/main/java/internal/sdmxdl/NoOpCaching.java
@@ -6,7 +6,6 @@ import sdmxdl.DataRepository;
 import sdmxdl.ErrorListener;
 import sdmxdl.EventListener;
 import sdmxdl.ext.Cache;
-import sdmxdl.ext.Persistence;
 import sdmxdl.file.FileSource;
 import sdmxdl.file.spi.FileCaching;
 import sdmxdl.web.Credentials;
@@ -16,7 +15,6 @@ import sdmxdl.web.spi.WebCaching;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
 
 public enum NoOpCaching implements FileCaching, WebCaching {
 
@@ -45,7 +43,6 @@ public enum NoOpCaching implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<DataRepository> getReaderCache(
             @NonNull FileSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return Cache.noOp();
@@ -54,7 +51,6 @@ public enum NoOpCaching implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<DataRepository> getDriverCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return Cache.noOp();
@@ -63,7 +59,6 @@ public enum NoOpCaching implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<MonitorReports> getMonitorCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return Cache.noOp();