+++ sdmx-dl-format-base/src/main/java/sdmxdl/format/caching/DiskCachingSupport.java
@@ -2,24 +2,29 @@ package sdmxdl.format.caching;
 
 import lombok.AccessLevel;
 import lombok.NonNull;
+import nbbrd.io.text.BooleanProperty;
+import nbbrd.io.text.Formatter;
+import nbbrd.io.text.Parser;
+import nbbrd.io.text.Property;
 import org.jspecify.annotations.Nullable;
 import sdmxdl.*;
 import sdmxdl.ext.Cache;
 import sdmxdl.ext.FileFormat;
-import sdmxdl.ext.Persistence;
 import sdmxdl.file.FileSource;
 import sdmxdl.file.spi.FileCaching;
+import sdmxdl.format.PropertiesSupport;
+import sdmxdl.format.design.PropertyDefinition;
 import sdmxdl.format.design.ServiceSupport;
 import sdmxdl.web.Credentials;
 import sdmxdl.web.MonitorReports;
 import sdmxdl.web.WebSource;
 import sdmxdl.web.spi.WebCaching;
 
+import java.io.File;
 import java.nio.file.Path;
 import java.time.Clock;
 import java.util.Collection;
-import java.util.List;
-import java.util.function.Predicate;
+import java.util.function.Function;
 
 import static java.util.Collections.emptyList;
 import static sdmxdl.format.FileFormatSupport.*;
@@ -29,6 +34,16 @@ import static sdmxdl.format.FileFormatSupport.*;
 @lombok.Builder(toBuilder = true)
 public final class DiskCachingSupport implements FileCaching, WebCaching {
 
+    // Set cache folder
+    @PropertyDefinition
+    public static final Property<File> CACHE_FOLDER_PROPERTY
+            = Property.of("sdmxdl.caching.cacheFolder", null, Parser.onFile(), Formatter.onFile());
+
+    // Disable cache compression
+    @PropertyDefinition
+    public static final BooleanProperty NO_COMPRESSION_PROPERTY
+            = BooleanProperty.of("sdmxdl.caching.noCompression", false);
+
     @lombok.Getter(AccessLevel.PRIVATE)
     private final @NonNull String id;
 
@@ -36,16 +51,13 @@ public final class DiskCachingSupport implements FileCaching, WebCaching {
     private final int rank = UNKNOWN_WEB_CACHING_RANK;
 
     @lombok.Builder.Default
-    private final @NonNull Path root = DiskCache.SDMXDL_TMP_DIR;
-
-    @lombok.Builder.Default
     private final @NonNull Clock clock = Clock.systemDefaultZone();
 
     @lombok.Builder.Default
-    private final boolean noCompression = false;
+    private final @NonNull FileFormat<DataRepository> repositoryFormat = FileFormat.noOp();
 
     @lombok.Builder.Default
-    private final @NonNull String persistenceId = "";
+    private final @NonNull FileFormat<MonitorReports> monitorsFormat = FileFormat.noOp();
 
     @Override
     public @NonNull String getFileCachingId() {
@@ -68,47 +80,65 @@ public final class DiskCachingSupport implements FileCaching, WebCaching {
     }
 
     @Override
-    public @NonNull Cache<DataRepository> getReaderCache(@NonNull FileSource source, @NonNull List<Persistence> persistences, @Nullable EventListener onEvent, @Nullable ErrorListener onError) {
-        FileFormat<DataRepository> repository = lookupFileFormat(DataRepository.class, persistences);
-        logConfig(onEvent, repository);
+    public @NonNull Cache<DataRepository> getReaderCache(
+            @NonNull FileSource source,
+            @Nullable EventListener onEvent,
+            @Nullable ErrorListener onError) {
+        Function<? super String, ? extends CharSequence> properties = PropertiesSupport.asFunction(source);
+        Path root = getCacheFolder(properties);
+
+        logConfig(onEvent, repositoryFormat, root);
         return decorateCache(DiskCache
                 .<DataRepository>builder()
+                .id(id)
                 .root(root)
-                .format(decorateFormat(repository))
+                .format(decorateFormat(repositoryFormat, isNoCompression(properties)))
                 .namePrefix("R")
                 .clock(clock)
-                .onRead(onEvent != null ? onEvent.asConsumer(id) : null)
-                .onError(onError != null ? onError.asBiConsumer(id) : null)
+                .onEvent(onEvent)
+                .onError(onError)
                 .build());
     }
 
     @Override
-    public @NonNull Cache<DataRepository> getDriverCache(@NonNull WebSource source, @NonNull List<Persistence> persistences, @Nullable EventListener onEvent, @Nullable ErrorListener onError) {
-        FileFormat<DataRepository> repository = lookupFileFormat(DataRepository.class, persistences);
-        logConfig(onEvent, repository);
+    public @NonNull Cache<DataRepository> getDriverCache(
+            @NonNull WebSource source,
+            @Nullable EventListener onEvent,
+            @Nullable ErrorListener onError) {
+        Function<? super String, ? extends CharSequence> properties = PropertiesSupport.asFunction(source);
+        Path root = getCacheFolder(properties);
+
+        logConfig(onEvent, repositoryFormat, root);
         return decorateCache(DiskCache
                 .<DataRepository>builder()
+                .id(id)
                 .root(root)
-                .format(decorateFormat(repository))
+                .format(decorateFormat(repositoryFormat, isNoCompression(properties)))
                 .namePrefix("D")
                 .clock(clock)
-                .onRead(onEvent != null ? onEvent.asConsumer(id) : null)
-                .onError(onError != null ? onError.asBiConsumer(id) : null)
+                .onEvent(onEvent)
+                .onError(onError)
                 .build());
     }
 
     @Override
-    public @NonNull Cache<MonitorReports> getMonitorCache(@NonNull WebSource source, @NonNull List<Persistence> persistences, @Nullable EventListener onEvent, @Nullable ErrorListener onError) {
-        FileFormat<MonitorReports> monitor = lookupFileFormat(MonitorReports.class, persistences);
-        logConfig(onEvent, monitor);
+    public @NonNull Cache<MonitorReports> getMonitorCache(
+            @NonNull WebSource source,
+            @Nullable EventListener onEvent,
+            @Nullable ErrorListener onError) {
+        Function<? super String, ? extends CharSequence> properties = PropertiesSupport.asFunction(source);
+        Path root = getCacheFolder(properties);
+
+        logConfig(onEvent, monitorsFormat, root);
         return decorateCache(DiskCache
                 .<MonitorReports>builder()
+                .id(id)
                 .root(root)
-                .format(decorateFormat(monitor))
+                .format(decorateFormat(monitorsFormat, isNoCompression(properties)))
                 .namePrefix("M")
                 .clock(clock)
-                .onRead(onEvent != null ? onEvent.asConsumer(id) : null)
-                .onError(onError != null ? onError.asBiConsumer(id) : null)
+                .onEvent(onEvent)
+                .onError(onError)
                 .build());
     }
 
@@ -127,22 +157,7 @@ public final class DiskCachingSupport implements FileCaching, WebCaching {
         return emptyList();
     }
 
-    private <T extends HasPersistence> FileFormat<T> lookupFileFormat(Class<T> type, List<Persistence> persistences) {
-        return persistences
-                .stream()
-                .filter(getPersistenceFilter(type))
-                .map(persistence -> persistence.getFormat(type))
-                .findFirst()
-                .orElseGet(FileFormat::noOp);
-    }
-
-    private Predicate<Persistence> getPersistenceFilter(Class<? extends HasPersistence> type) {
-        return persistenceId.isEmpty()
-                ? (persistence -> persistence.getFormatSupportedTypes().contains(type))
-                : (persistence -> persistence.getPersistenceId().equals(persistenceId));
-    }
-
-    private <T extends HasPersistence> FileFormat<T> decorateFormat(FileFormat<T> format) {
+    private <T extends HasPersistence> FileFormat<T> decorateFormat(FileFormat<T> format, boolean noCompression) {
         return lock(noCompression ? wrap(format) : gzip(wrap(format)));
     }
 
@@ -159,8 +174,17 @@ public final class DiskCachingSupport implements FileCaching, WebCaching {
                 .build();
     }
 
-    private void logConfig(EventListener onEvent, FileFormat<?> format) {
+    private void logConfig(EventListener onEvent, FileFormat<?> format, Path root) {
         if (onEvent != null)
             onEvent.accept(id, "Using cache folder " + root.toUri() + " with format '" + format.getFileExtension() + "'");
+    }
+
+    private static Path getCacheFolder(Function<? super String, ? extends CharSequence> properties) {
+        File root = CACHE_FOLDER_PROPERTY.get(properties);
+        return (root == null) ? DiskCache.SDMXDL_TMP_DIR : root.toPath();
+    }
+
+    private static boolean isNoCompression(Function<? super String, ? extends CharSequence> properties) {
+        return NO_COMPRESSION_PROPERTY.get(properties);
     }
 }