+++ sdmx-dl-format-base/src/main/java/sdmxdl/format/caching/MemCachingSupport.java
@@ -6,7 +6,6 @@ import sdmxdl.DataRepository;
 import sdmxdl.ErrorListener;
 import sdmxdl.EventListener;
 import sdmxdl.ext.Cache;
-import sdmxdl.ext.Persistence;
 import sdmxdl.file.FileSource;
 import sdmxdl.file.spi.FileCaching;
 import sdmxdl.format.design.ServiceSupport;
@@ -16,7 +15,10 @@ import sdmxdl.web.WebSource;
 import sdmxdl.web.spi.WebCaching;
 
 import java.time.Clock;
-import java.util.*;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.ConcurrentMap;
 import java.util.function.Supplier;
 
@@ -70,7 +72,6 @@ public final class MemCachingSupport implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<DataRepository> getReaderCache(
             @NonNull FileSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return MemCache
@@ -83,7 +84,6 @@ public final class MemCachingSupport implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<DataRepository> getDriverCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return MemCache
@@ -96,7 +96,6 @@ public final class MemCachingSupport implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<MonitorReports> getMonitorCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
         return MemCache