+++ sdmx-dl-provider-ri/src/main/java/sdmxdl/provider/ri/caching/RiCaching.java
@@ -11,24 +11,20 @@ import nbbrd.service.ServiceProvider;
 import org.jspecify.annotations.Nullable;
 import sdmxdl.*;
 import sdmxdl.ext.Cache;
-import sdmxdl.ext.Persistence;
 import sdmxdl.file.FileSource;
 import sdmxdl.file.spi.FileCaching;
-import sdmxdl.format.caching.DiskCache;
+import sdmxdl.format.PropertiesSupport;
 import sdmxdl.format.caching.DiskCachingSupport;
 import sdmxdl.format.caching.MemCache;
 import sdmxdl.format.design.PropertyDefinition;
-import sdmxdl.provider.PropertiesSupport;
+import sdmxdl.format.kryo.KryoFileFormat;
 import sdmxdl.web.Credentials;
 import sdmxdl.web.MonitorReports;
 import sdmxdl.web.WebSource;
 import sdmxdl.web.spi.WebCaching;
 
-import java.io.File;
-import java.nio.file.Path;
 import java.time.Duration;
 import java.util.Collection;
-import java.util.List;
 import java.util.function.Function;
 
 import static java.util.Objects.requireNonNull;
@@ -44,21 +40,6 @@ public final class RiCaching implements FileCaching, WebCaching {
     public static final BooleanProperty NO_CACHE_PROPERTY
             = BooleanProperty.of("sdmxdl.caching.noCache", false);
 
-    // Set cache folder
-    @PropertyDefinition
-    public static final Property<File> CACHE_FOLDER_PROPERTY
-            = Property.of("sdmxdl.caching.cacheFolder", null, Parser.onFile(), Formatter.onFile());
-
-    // Disable cache compression
-    @PropertyDefinition
-    public static final BooleanProperty NO_COMPRESSION_PROPERTY
-            = BooleanProperty.of("sdmxdl.caching.noCompression", false);
-
-    // Set cache persistence backend
-    @PropertyDefinition
-    public static final Property<String> PERSISTENCE_ID_PROPERTY
-            = Property.of("sdmxdl.caching.persistenceId", null, Parser.onString(), Formatter.onString());
-
     // Set max confidentiality
     @PropertyDefinition
     public static final Property<Confidentiality> MAX_CONFIDENTIALITY_PROPERTY
@@ -70,13 +51,23 @@ public final class RiCaching implements FileCaching, WebCaching {
 
     private static final Collection<String> PROPERTIES = keysOf(
             NO_CACHE_PROPERTY,
-            CACHE_FOLDER_PROPERTY,
-            NO_COMPRESSION_PROPERTY,
-            PERSISTENCE_ID_PROPERTY,
+            DiskCachingSupport.CACHE_FOLDER_PROPERTY,
+            DiskCachingSupport.NO_COMPRESSION_PROPERTY,
             MAX_CONFIDENTIALITY_PROPERTY
     );
 
-    private final VaultCachingSupport vaultCaching = VaultCachingSupport.builder().id(ID).build();
+    private final DiskCachingSupport diskCaching = DiskCachingSupport
+            .builder()
+            .id(ID)
+            .rank(RANK)
+            .repositoryFormat(KryoFileFormat.of(DataRepository.class))
+            .monitorsFormat(KryoFileFormat.of(MonitorReports.class))
+            .build();
+
+    private final VaultCachingSupport vaultCaching = VaultCachingSupport
+            .builder()
+            .id(ID)
+            .build();
 
     @Override
     public @NonNull String getWebCachingId() {
@@ -111,7 +102,6 @@ public final class RiCaching implements FileCaching, WebCaching {
     @Override
     public @NonNull Cache<DataRepository> getDriverCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
 
@@ -125,14 +115,13 @@ public final class RiCaching implements FileCaching, WebCaching {
             return forbiddenCache(onEvent);
         }
 
-        return getDiskCaching(properties)
-                .getDriverCache(source, persistences, onEvent, onError);
+        return diskCaching
+                .getDriverCache(source, onEvent, onError);
     }
 
     @Override
     public @NonNull Cache<MonitorReports> getMonitorCache(
             @NonNull WebSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
 
@@ -146,8 +135,8 @@ public final class RiCaching implements FileCaching, WebCaching {
             return forbiddenCache(onEvent);
         }
 
-        return getDiskCaching(properties)
-                .getMonitorCache(source, persistences, onEvent, onError);
+        return diskCaching
+                .getMonitorCache(source, onEvent, onError);
     }
 
     @Override
@@ -155,13 +144,14 @@ public final class RiCaching implements FileCaching, WebCaching {
             @NonNull WebSource source,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
-        return vaultCaching.getCredentialsCache(Duration.ofMinutes(5), onEvent, onError);
+
+        return vaultCaching
+                .getCredentialsCache(Duration.ofMinutes(5), onEvent, onError);
     }
 
     @Override
     public @NonNull Cache<DataRepository> getReaderCache(
             @NonNull FileSource source,
-            @NonNull List<Persistence> persistences,
             @Nullable EventListener onEvent,
             @Nullable ErrorListener onError) {
 
@@ -171,8 +161,8 @@ public final class RiCaching implements FileCaching, WebCaching {
             return noCache(onEvent);
         }
 
-        return getDiskCaching(properties)
-                .getReaderCache(source, persistences, onEvent, onError);
+        return diskCaching
+                .getReaderCache(source, onEvent, onError);
     }
 
     private <V extends HasExpiration> Cache<V> noCache(EventListener onEvent) {
@@ -185,17 +175,6 @@ public final class RiCaching implements FileCaching, WebCaching {
         return MemCache.<V>builder().build();
     }
 
-    private static DiskCachingSupport getDiskCaching(Function<? super String, ? extends CharSequence> properties) {
-        return DiskCachingSupport
-                .builder()
-                .id(ID)
-                .rank(RANK)
-                .root(getCacheFolder(properties))
-                .noCompression(isNoCompression(properties))
-                .persistenceId(getPersistenceId(properties))
-                .build();
-    }
-
     private static boolean isNoCache(Function<? super String, ? extends CharSequence> properties) {
         return NO_CACHE_PROPERTY.get(properties);
     }
@@ -203,19 +182,5 @@ public final class RiCaching implements FileCaching, WebCaching {
     @VisibleForTesting
     static boolean isForbidden(Function<? super String, ? extends CharSequence> properties, Confidentiality confidentiality) {
         return confidentiality.compareTo(requireNonNull(MAX_CONFIDENTIALITY_PROPERTY.get(properties))) > 0;
-    }
-
-    private static Path getCacheFolder(Function<? super String, ? extends CharSequence> properties) {
-        File root = CACHE_FOLDER_PROPERTY.get(properties);
-        return (root == null) ? DiskCache.SDMXDL_TMP_DIR : root.toPath();
-    }
-
-    private static boolean isNoCompression(Function<? super String, ? extends CharSequence> properties) {
-        return NO_COMPRESSION_PROPERTY.get(properties);
-    }
-
-    private static String getPersistenceId(Function<? super String, ? extends CharSequence> properties) {
-        String result = PERSISTENCE_ID_PROPERTY.get(properties);
-        return result != null ? result : "";
     }
 }